Re-apply changes to test_sampler.
